+++ script/main.js
@@ -18,8 +18,26 @@ let script = (function () {
     });
   };
 
+  let deleteSnippet = function (path, snippetName, rowIndex) {
+    vscode.postMessage({
+      command: "deleteSnippet",
+      path,
+      snippetName,
+    });
+
+    let decodedPath = decodeURIComponent(path);
+    let row = document.querySelector(
+      `table[data-path='${decodedPath}'] tbody tr:nth-child(${rowIndex})`
+    );
+
+    if (row !== null) {
+      row.remove();
+    }
+  };
+
   return {
     openFile,
     editSnippet,
+    deleteSnippet,
   };
 })();