+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] - 2021-04-23
+
+### Added
+
+- Delete action.
+
+### Changed
+
+- Edit action to open snippet with the cursor at the first character of the snippet name.
+- Improved appearance at different viewport sizes.
+
 ## [0.16.0] - 2021-04-22
 
 ### Changed